fix: avoid unnecessary typescript checks in /files when building (#757) (#766)
