+++ explorer/app/apis/catalog/common/entities.ts
@@ -1,0 +1,11 @@
+export interface DbGapStudy {
+  consentCodes: string[];
+  dataTypes: string[];
+  dbGapId: string;
+  description: string;
+  focus: string;
+  participantCount: number;
+  studyAccession: string;
+  studyDesigns: string[];
+  title: string;
+}
