+++ explorer/app/apis/catalog/ncpi-catalog/common/entities.ts
@@ -1,3 +1,16 @@
+import { DbGapStudy } from "../../common/entities";
+
+export interface PlatformStudy {
+  dbGapId: string;
+  platform: string;
+}
+
+export interface NCPIStudy extends DbGapStudy {
+  platforms: string[];
+}
+
+export type DbGapId = string;
+
 export type NCPICatalogEntity = NCPICatalogPlatform | NCPICatalogStudy;
 
 export interface NCPICatalogPlatform {
