+++ explorer/files/ncpi-catalog/build-platform-studies.ts
@@ -1,5 +1,9 @@
-import { DbGapStudy, getStudy } from "../common/dbGaP";
-import { NCPIStudy, PlatformStudy } from "./entities";
+import { DbGapStudy } from "../../app/apis/catalog/common/entities";
+import {
+  NCPIStudy,
+  PlatformStudy,
+} from "../../app/apis/catalog/ncpi-catalog/common/entities";
+import { getStudy } from "../common/dbGaP";
 
 /**
  * Build the catalog platform studies for NCPI.
